Move pylxd.tests to pylxd.deprecated.tests
