+++ pylxd/deprecated/tests/fake_api.py
