+++ pylxd/deprecated/tests/test_client.py
