+++ pylxd/deprecated/tests/test_connection.py
@@ -23,7 +23,7 @@ import unittest
 
 from pylxd.deprecated import connection
 from pylxd.deprecated import exceptions
-from pylxd.tests import annotated_data
+from pylxd.deprecated.tests import annotated_data
 
 if six.PY34:
     from io import BytesIO
