+++ pylxd/deprecated/tests/test_container.py
@@ -20,9 +20,9 @@ import tempfile
 
 from pylxd.deprecated import connection
 
-from pylxd.tests import annotated_data
-from pylxd.tests import fake_api
-from pylxd.tests import LXDAPITestBase
+from pylxd.deprecated.tests import annotated_data
+from pylxd.deprecated.tests import fake_api
+from pylxd.deprecated.tests import LXDAPITestBase
 
 
 @ddt
