+++ pylxd/deprecated/tests/test_image.py
@@ -23,9 +23,9 @@ from pylxd.deprecated import connection
 from pylxd.deprecated import exceptions
 from pylxd.deprecated import image
 
-from pylxd.tests import annotated_data
-from pylxd.tests import fake_api
-from pylxd.tests import LXDAPITestBase
+from pylxd.deprecated.tests import annotated_data
+from pylxd.deprecated.tests import fake_api
+from pylxd.deprecated.tests import LXDAPITestBase
 
 
 @ddt
