+++ pylxd/deprecated/tests/test_operation.py
@@ -18,9 +18,9 @@ import mock
 
 from pylxd.deprecated import connection
 
-from pylxd.tests import annotated_data
-from pylxd.tests import fake_api
-from pylxd.tests import LXDAPITestBase
+from pylxd.deprecated.tests import annotated_data
+from pylxd.deprecated.tests import fake_api
+from pylxd.deprecated.tests import LXDAPITestBase
 
 
 @ddt
